NIFI-5775 DataTypeUtils "toString" incorrectly treats value as a "byte" when passing an array leading to ClassCastException. If it is an array of objects, it is converted to String by iterating over the entire object and converting it to string and appending to variable sb which is a Stringbuilder.
NIFI-5775 DataTypeUtils "toString" incorrectly treats value as a "byte" when passing an array leading to ClassCastException. If it is an array of objects, it is converted to String by iterating over the entire object and converting it to string and appending to variable sb which is a Stringbuilder.

NIFI-5775 DataTypeUtils "toString" incorrectly treats value as a "byte" when passing an array leading to ClassCastException. Adding of new test testChoiceArray1 to TestWriteJsonResult class.